Replace u32 with usize
For non-STV methods, this is very slightly slower for few voters,
and very slightly faster for more voters, on a 64-bit machine,
because usize is essentially u64. The differences are not
noticeable though, and probably not statistically significant.
The benefits is better readability and maintenance,
which outweighs the negligible performance hit.

For STV, this is very slightly faster in all 6 situations
tested in the STV config.

+++ rust/config.rs
@@ -31,7 +31,7 @@ impl AllocationMethod {
     }
     pub fn simulate_elections(
         &self,
-        n_seats: u32,
+        n_seats: usize,
         n_voters: usize,
         parties: &[Party],
         bar: &Option<ProgressBar>,
@@ -88,7 +88,7 @@ pub struct Config {
     pub data_out_dir: String,
 
     /// Total number of seats in this district (district magnitude)
-    pub n_seats: u32,
+    pub n_seats: usize,
 
     /// Total number of voters to generate with the normal distribution
     pub n_voters: usize,
